devices: move vtl permissions guard to its own crate
