+++ vm/devices/vmbus/vmbus_relay_intercept_device/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 rust-version.workspace = true
 
 [dependencies]
+lower_vtl_permissions_guard.workspace = true
 vmbus_channel.workspace = true
 vmbus_client.workspace = true
 vmbus_core.workspace = true
