+++ vm/devices/vmbus/vmbus_relay_intercept_device/src/lib.rs
@@ -18,8 +18,8 @@ use anyhow::Context;
 use anyhow::Result;
 use futures::StreamExt;
 use guid::Guid;
-use inspect::Inspect;
 use inspect::InspectMut;
+use lower_vtl_permissions_guard::PagesAccessibleToLowerVtl;
 use mesh::rpc::RpcSend;
 use pal_async::driver::SpawnDriver;
 use std::future::Future;
@@ -637,51 +637,51 @@ impl SignalVmbusChannel for MemoryBlockChannelSignal {
     }
 }
 
-#[derive(Inspect)]
-struct PagesAccessibleToLowerVtl {
-    #[inspect(skip)]
-    vtl_protect: Arc<dyn VtlMemoryProtection + Send + Sync>,
-    #[inspect(with = "|x| inspect::iter_by_index(x).map_value(inspect::AsHex)")]
-    pages: Vec<u64>,
-}
+// #[derive(Inspect)]
+// struct PagesAccessibleToLowerVtl {
+//     #[inspect(skip)]
+//     vtl_protect: Arc<dyn VtlMemoryProtection + Send + Sync>,
+//     #[inspect(with = "|x| inspect::iter_by_index(x).map_value(inspect::AsHex)")]
+//     pages: Vec<u64>,
+// }
 
-impl PagesAccessibleToLowerVtl {
-    pub fn new_from_memory_block(
-        vtl_protect: Arc<dyn VtlMemoryProtection + Send + Sync>,
-        memory: &MemoryBlock,
-    ) -> Result<Self> {
-        let pages = Vec::with_capacity(memory.pfns().len());
-        let mut this = Self { vtl_protect, pages };
-        for pfn in memory.pfns() {
-            this.vtl_protect
-                .modify_vtl_page_setting(*pfn, hvdef::HV_MAP_GPA_PERMISSIONS_ALL)
-                .context("failed to update VTL protections on page")?;
-            this.pages.push(*pfn);
-        }
-        Ok(this)
-    }
+// impl PagesAccessibleToLowerVtl {
+//     pub fn new_from_memory_block(
+//         vtl_protect: Arc<dyn VtlMemoryProtection + Send + Sync>,
+//         memory: &MemoryBlock,
+//     ) -> Result<Self> {
+//         let pages = Vec::with_capacity(memory.pfns().len());
+//         let mut this = Self { vtl_protect, pages };
+//         for pfn in memory.pfns() {
+//             this.vtl_protect
+//                 .modify_vtl_page_setting(*pfn, hvdef::HV_MAP_GPA_PERMISSIONS_ALL)
+//                 .context("failed to update VTL protections on page")?;
+//             this.pages.push(*pfn);
+//         }
+//         Ok(this)
+//     }
 
-    pub fn pfns(&self) -> &[u64] {
-        &self.pages
-    }
-}
+//     pub fn pfns(&self) -> &[u64] {
+//         &self.pages
+//     }
+// }
 
-impl Drop for PagesAccessibleToLowerVtl {
-    fn drop(&mut self) {
-        if let Err(err) = self
-            .pages
-            .iter()
-            .map(|pfn| {
-                self.vtl_protect
-                    .modify_vtl_page_setting(*pfn, hvdef::HV_MAP_GPA_PERMISSIONS_NONE)
-                    .context("failed to update VTL protections on page")
-            })
-            .collect::<Result<Vec<_>>>()
-        {
-            tracing::error!(
-                err = err.as_ref() as &dyn std::error::Error,
-                "Failed resetting page protections"
-            );
-        }
-    }
-}
+// impl Drop for PagesAccessibleToLowerVtl {
+//     fn drop(&mut self) {
+//         if let Err(err) = self
+//             .pages
+//             .iter()
+//             .map(|pfn| {
+//                 self.vtl_protect
+//                     .modify_vtl_page_setting(*pfn, hvdef::HV_MAP_GPA_PERMISSIONS_NONE)
+//                     .context("failed to update VTL protections on page")
+//             })
+//             .collect::<Result<Vec<_>>>()
+//         {
+//             tracing::error!(
+//                 err = err.as_ref() as &dyn std::error::Error,
+//                 "Failed resetting page protections"
+//             );
+//         }
+//     }
+// }
